# backend/docs/swagger.yaml
basePath: /
definitions:
...
  handler.Restaurant:
    properties:
      address:
        type: string
      id:
        type: integer
      logo_url:
        type: string
      name:
        type: string
      phone_number:
        type: string
      user_id:
        type: integer
    type: object
...
host: localhost:3000
info:
  contact: {}
  description: API for Order System with user authentication and restaurant management
  title: Order System API
  version: "1.0"
paths:
...
  /api/restaurant/:
    get:
      description: Get all restaurants for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handler.Restaurant'
            type: array
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Error retrieving restaurants
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get all restaurants
      tags:
      - Restaurant
    post:
      consumes:
      - application/json
      description: Create a new restaurant for the authenticated user
      parameters:
      - description: Restaurant data
        in: body
        name: restaurant
        required: true
        schema:
          $ref: '#/definitions/handler.Restaurant'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handler.Restaurant'
        "400":
          description: Invalid input
          schema:
            type: string
        "404":
          description: User not found
          schema:
            type: string
        "500":
          description: Error creating restaurant
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create a new restaurant
      tags:
      - Restaurant
  /api/restaurant/{id}:
    delete:
      description: Delete a restaurant by ID
      parameters:
      - description: Restaurant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "404":
          description: User or restaurant not found
          schema:
            type: string
        "500":
          description: Error deleting restaurant
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete restaurant
      tags:
      - Restaurant
    get:
      description: Get a restaurant by ID without authentication
      parameters:
      - description: Restaurant ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Restaurant'
        "404":
          description: Restaurant not found
          schema:
            type: string
      summary: Get public restaurant by ID
      tags:
      - Restaurant
    put:
      consumes:
      - application/json
      description: Update a restaurant by ID
      parameters:
      - description: Restaurant ID
        in: path
        name: id
        required: true
        type: string
      - description: Restaurant data
        in: body
        name: restaurant
        required: true
        schema:
          $ref: '#/definitions/handler.Restaurant'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.Restaurant'
        "400":
          description: Invalid input
          schema:
            type: string
        "404":
          description: User or restaurant not found
          schema:
            type: string
        "500":
          description: Error updating restaurant
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update restaurant
      tags:
      - Restaurant
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"